refactor: tests changed for 0.34.1 of sn_api
With our new versioning scheme that uses the VersionHash rather than
integers, a couple of the `cli_cat` tests had to be updated to take
that into account. At the same time, those test cases were updated to
use assert_fs for the test data that was provided.

The various tests that were using NRS and versioning were updated to
incorporate the new VersionHash scheme. For each test that had to be
modified, it was updated to use `assert_fs` for the test data.

There was also a sync test that was changed to use `assert_fs` because
it was deleting a file in the test data. This would then produce
unexpected results on subsequent test runs, as other tests were
expecting that data to exist.

As of this commit, all of these tests are passing individually. The
frequency of 'Failed to GET Public Blob' failures in the full test run
are drastically reduced with the newer versiob of the network.

There were lots of tests that were explicitly using version 0 in their
source for the `files get` commands. There was also a test that was
dealing with an NRS link that was modified for new behaviour.