feat: updates for new sn_api and safe_network 0.31.x
Version 0.31.x of `safe_network` and in turn `sn_api` has changes for connecting to the network.
Clients who connect now need to supply what we're calling the 'genesis key', which is a BLS public
key. The API was updated for this and it has a significant impact on the CLI, mainly around the
`network` commands, which need to work with this new configuration.