feat: updates for new sn_api and safe_network 0.31.x
Version 0.31.x of `safe_network` and in turn `sn_api` has changes for connecting to the network.
Clients who connect now need to supply what we're calling the 'genesis key', which is a BLS public
key. The API was updated for this and it has a significant impact on the CLI, mainly around the
`network` commands, which need to work with this new configuration.

Changes to support this new version:

* Self encryption no longer permitted files with 0 bytes, and a couple of files in our test data
  have 0 bytes, so I just added some comments to these.
* The `xor_name` crate was upgraded to use the same version as `safe_network`.
* The `bls` crate was upgraded to use the same version as `safe_network`.
* The `bytes` crate was added to work with the `safe_network` and `sn_api` which are now using the
  `Bytes` type rather than slices of `u8`.