chore: change logging framework to tracing crate
The `safe_network` repository had been updated to use the `tracing` crate for logging, rather than
the `env_logger` crate, because tracing offers more options. It turned out that this wasn't
compatible with the CLI, which was still using the `env_logger` crate, and this meant you didn't get
logging output in `safe` from any of the client side code in `safe_network`.

Right now, whether logging is enabled is just done by checking the `RUST_LOG` variable. If it's not
set, the user will just get the value of the println output. If it is set, the logging to be done
will be determined by the value of that variable.

We now get the `safe_network` output.