chore: force new release with readme update
BREAKING CHANGE: Multimap entries for NRS will be assigned a different data type when they are
created.

Note, the breaking change relates to the previous commit. For some reason the workflow for updating
the version and deploying a new release didn't trigger.

This commit and PR is just to try and force a new release.