ci: run test suite against specific node version
For now the CLI works best against a specific version of `safe_network` and it doesn't necessarily
make sense to test it against the latest version.

The test for the Bash install script has been removed for Windows since this isn't really a valid
test on that platform. We really need to get a Powershell script for installation on Windows.

The node installation tests have been removed. I'm going to write full integration or unit tests for
this in Rust, as that means the tests will be cross platform, so we won't have to do something
different for Windows and Unix systems.